restructured project folder with new files
